Add Hive support to NoteModel
This commit adds Hive support to the NoteModel class by creating a HiveObject
and generating a TypeAdapter for it. The NoteModel class now has HiveFields for
each of its properties, allowing it to be stored in a Hive box. The command to
generate the TypeAdapter is also included in a comment at the end of the file.