Add Hive database initialization
This commit adds initialization of the Hive database by opening the 'notes box' box. The 'initHive' function is called in the main function to ensure that the database is initialized before the app starts. The 'const.dart' file is also added to store constant values, including the name of the Hive box.