Refactor code to use a separate AddNoteCubit for adding notes
This commit refactors the code to use a separate AddNoteCubit for adding notes instead of using the MainCubit. The AddNoteCubit is responsible for validating and adding notes to the Hive database. The AddNoteBottomSheet and AddNoteForm widgets have been updated to use the new AddNoteCubit. The MainCubit has been commented out for now. Additionally, a SimpleBlocObserver has been added to observe BLoC events.